Extract crypto methods into seperate service
